+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -249,6 +249,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`hooks-slot-limit
+  <clang-tidy/checks/hooks-slot-limit>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`hooks-validate-buf-len
   <clang-tidy/checks/hooks-validate-buf-len>` check.